Add support for encoded passwords on config files
Passwords can now be encoded as Base64 or using the Jetty
password encoding tool.

The runtime then utilizes the password prefix OBF: or B64:
to identify the algorithm to use when decoding the password
to plain text.

If the password has no prefix, the provided value is returned
as is.